+++ packages/main/src/RPA/Email/Exchange.py
@@ -103,14 +103,24 @@ class Exchange(OAuthMixin):
 
         *** Settings ***
         Library     RPA.Email.Exchange
-        Task Setup  Authorize  username=${ACCOUNT}  password=${PASSWORD}
+        ...     vault_name=email_oauth_microsoft    vault_token_key=token
+        ...     tenant=ztzvn.onmicrosoft.com  # your custom tenant here
+        Task Setup      Ensure Auth
 
         *** Variables ***
         ${ACCOUNT}              ACCOUNT_NAME
-        ${PASSWORD}             ACCOUNT_PASSWORD
         ${RECIPIENT_ADDRESS}    RECIPIENT
         ${IMAGES}               myimage.png
         ${ATTACHMENTS}          C:${/}files${/}mydocument.pdf
+
+        *** Keywords ***
+        Ensure Auth
+            ${secrets} =    Get Secret    email_oauth_microsoft
+            RPA.Email.Exchange.Authorize    ${ACCOUNT}
+            ...    is_oauth=${True}  # use the OAuth2 auth code flow (required)
+            ...    client_id=${secrets}[client_id]  # app ID
+            ...    client_secret=${secrets}[client_secret]  # app password
+            ...    token=${secrets}[token]  # token dict (access, refresh, scope etc.)
 
         *** Tasks ***
         Task of sending email
@@ -153,12 +163,24 @@ class Exchange(OAuthMixin):
     .. code-block:: python
 
         from RPA.Email.Exchange import Exchange
+        from RPA.Robocorp.Vault import Vault
 
+        vault_name = "email_oauth_microsoft"
+        secrets = Vault().get_secret(vault_name)
         ex_account = "ACCOUNT_NAME"
-        ex_password = "ACCOUNT_PASSWORD"
 
-        mail = Exchange()
-        mail.authorize(username=ex_account, password=ex_password)
+        mail = Exchange(
+            vault_name=vault_name,
+            vault_token_key="token",
+            tenant="ztzvn.onmicrosoft.com"
+        )
+        mail.authorize(
+            username=ex_account,
+            is_oauth=True,
+            client_id=secrets["client_id"],
+            client_secret=secrets["client_secret"],
+            token=secrets["token"]
+        )
         mail.send_message(
             recipients="RECIPIENT",
             subject="Message from RPA Python",
@@ -167,21 +189,14 @@ class Exchange(OAuthMixin):
 
     **OAuth2**
 
-    .. code-block:: robotframework
+    The OAuth2 flow is the only way of authorizing at the moment as Microsoft disabled
+    entirely the usage of passwords, even App Passwords. And since you have to work
+    with tokens now and because this library has the capability to automatically
+    refresh an expired token, please don't forget to initialize the library with the
+    following parameters: `vault_name`, `vault_token_key` and `tenant`.
 
-        *** Settings ***
-        Library     RPA.Email.Exchange
-        ...     vault_name=email_oauth_microsoft    vault_token_key=token
-
-        Task Setup      Authorize   username=${ACCOUNT}
-        ...    autodiscover=${False}    server=outlook.office365.com
-        ...    is_oauth=${True}  # use the OAuth2 auth code flow
-        # With secrets retrieved from the Vault by using ``Get Secret`` and set
-        #  globally into `${SECRETS}`.
-        ...    client_id=${SECRETS}[client_id]  # app ID
-        ...    client_secret=${SECRETS}[client_secret]  # app password
-        # The entire token structure auto refreshes when it expires.
-        ...    token=${SECRETS}[token]  # token dict (access, refresh, scopes etc.)
+    Learn more on how to use the OAuth2 flow in this Portal robot
+    `example-oauth-email <https://github.com/robocorp/example-oauth-email>`_.
 
     **About criterion parameter**